Added debug viz for physics. Stupid.
It was the crate after all. The model and the shape didn't match. Since
I copied it from the example, more or less, I didn't suspect it'd be
broken. Regardless, the debug viz shows what was wrong, and so, the
issue is solved.

Since the simple Bullet CC is functioning as expected, this is ok. But
I've also tested out @peterpodgorski 's KCC, and it's also just lovely,
as expected when all the work is done. I've added that example too.

Closes #3